+++ dbmodel/ud/ftrg/ud_gw_trg_edit_man_arc.sql
@@ -296,11 +296,12 @@ This version of Giswater is provided by Giswater Association
 				ELSIF (OLD.epa_type = 'VIRTUAL') THEN 
 				inp_table:= 'inp_virtual';
 				END IF;
-				v_sql:= 'DELETE FROM '||inp_table||' WHERE arc_id = '||quote_literal(OLD.arc_id);
-				EXECUTE v_sql;
-				
+				IF inp_table IS NOT NULL THEN
+					v_sql:= 'DELETE FROM '||inp_table||' WHERE arc_id = '||quote_literal(OLD.arc_id);
+					EXECUTE v_sql;
+				END IF;
+			
 				inp_table := NULL;
-
 
 				IF (NEW.epa_type = 'CONDUIT') THEN 
 				inp_table:= 'inp_conduit';
@@ -315,8 +316,10 @@ This version of Giswater is provided by Giswater Association
 				ELSIF (NEW.epa_type = 'VIRTUAL') THEN 
 				inp_table:= 'inp_virtual';
 				END IF;
-				v_sql:= 'INSERT INTO '||inp_table||' (arc_id) VALUES ('||quote_literal(NEW.arc_id)||')';
-				EXECUTE v_sql;
+				IF inp_table IS NOT NULL THEN
+					v_sql:= 'DELETE FROM '||inp_table||' WHERE arc_id = '||quote_literal(OLD.arc_id);
+					EXECUTE v_sql;
+				END IF;
 
 			END IF;
