+++ dbmodel/ud/ftrg/ud_gw_trg_edit_node.sql
@@ -226,8 +226,10 @@ BEGIN
             ELSIF (OLD.epa_type = 'STORAGE') THEN
                 inp_table:= 'inp_storage';    
 			END IF;
-            v_sql:= 'DELETE FROM '||inp_table||' WHERE node_id = '||quote_literal(OLD.node_id);
-            EXECUTE v_sql;
+			IF inp_table IS NOT NULL THEN
+                v_sql:= 'DELETE FROM '||inp_table||' WHERE node_id = '||quote_literal(OLD.node_id);
+                EXECUTE v_sql;
+            END IF;
 			inp_table := NULL;
 
             IF (NEW.epa_type = 'JUNCTION') THEN
@@ -239,8 +241,10 @@ BEGIN
             ELSIF (NEW.epa_type = 'STORAGE') THEN
                 inp_table:= 'inp_storage';
             END IF;
-            v_sql:= 'INSERT INTO '||inp_table||' (node_id) VALUES ('||quote_literal(NEW.node_id)||')';
-            EXECUTE v_sql;
+			IF inp_table IS NOT NULL THEN
+                v_sql:= 'DELETE FROM '||inp_table||' WHERE node_id = '||quote_literal(OLD.node_id);
+                EXECUTE v_sql;
+            END IF;
 
         END IF;
